Update Master_Contributor.md
President and CEO of AWS , API Autobahn Expressway platform  . 
We are interrupting are regular production lines to address a international crisis of copycat cases that seems to be very similar. But cases are stretched out . Or in a different state.  Because it possible because of ple bargain and null disclosure agreement the actual suspect got away with in . And only seems I've been so supported through my journey this is contribution of my apparentton